+++ shared/ui/dropdown/index.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { CSSProperties, ReactNode, createContext } from 'react'
+import { createContext } from 'react'
 
 import { CheckboxIcon, CloseIcon, OpenIcon } from '@/public/icons'
 import classNames from 'classnames/bind'
@@ -9,7 +9,7 @@ import { useDropdown } from './hooks/use-dropdown'
 import { useDropdownContext } from './hooks/use-dropdown-context'
 import RangeOption from './option/range-option'
 import styles from './styles.module.scss'
-import { DropdownItemProps, DropdownValueType } from './types'
+import { DropdownItemProps, DropdownProps } from './types'
 
 const cx = classNames.bind(styles)
 
@@ -28,17 +28,8 @@ const defaultContext = {
 
 export const DropdownContext = createContext<DropdownContextProps>(defaultContext)
 
-export interface DropdownProps {
-  Trigger: ReactNode
-  value: DropdownValueType
-  onChange: (value: DropdownValueType) => void
-  isMultiple: boolean
-  containerStyle?: CSSProperties
-  labelStyle?: CSSProperties
-  children?: ReactNode
-}
-
 const Dropdown = ({
+  size = 'small',
   Trigger,
   value,
   onChange,
@@ -55,31 +46,42 @@ const Dropdown = ({
 
   return (
     <DropdownContext.Provider value={{ isOpen, toggleOpen, handleSelect }}>
-      <div style={containerStyle} ref={dropdownRef}>
-        <button onClick={toggleOpen} className={cx('container', 'trigger')} style={labelStyle}>
+      <div className={cx(`dropdown-${size}`)} style={containerStyle} ref={dropdownRef}>
+        <button
+          onClick={toggleOpen}
+          className={cx('container', 'trigger', size, { open: isOpen })}
+          style={labelStyle}
+        >
           {Trigger}
           {isOpen ? <OpenIcon /> : <CloseIcon />}
         </button>
-        {isOpen && <ul className={cx('options')}>{children}</ul>}
+        {isOpen && <ul className={cx('options', size)}>{children}</ul>}
       </div>
     </DropdownContext.Provider>
   )
 }
 
-const Item = ({ isSelected, value, label, hasCheck = false, style }: DropdownItemProps) => {
+const Item = ({
+  isSelected,
+  value,
+  label,
+  hasCheck = false,
+  size = 'small',
+  style,
+}: DropdownItemProps) => {
   const context = useDropdownContext()
 
   const { handleSelect } = context
 
   return (
     <li
-      className={cx('container', 'option', { selected: isSelected })}
+      className={cx('container', 'option', size, { selected: isSelected })}
       style={style}
       onClick={() => handleSelect(value)}
       aria-hidden
     >
       {label}
-      {hasCheck && <CheckboxIcon className={cx('icon', { 'selected-icon': isSelected })} />}
+      {hasCheck && <CheckboxIcon className={cx({ 'selected-icon': isSelected })} />}
     </li>
   )
 }
