+++ shared/ui/dropdown/styles.module.scss
@@ -1,48 +1,107 @@
-// NOTE: color, padding 등 세부 수치들 미정, 일단 제일 가까운 걸로 함
+$small-width: 90px;
+$small-height: 30px;
+$large-width: 160px;
+$large-height: 40px;
+
+$color-selected-text: #171717;
+
+.dropdown {
+  &-small {
+    width: $small-width;
+    height: fit-content;
+  }
+
+  &-large {
+    width: $large-width;
+    height: fit-content;
+  }
+}
 
 .container {
-  width: 100%;
-  padding: 7px 11px;
-  border: 1px solid $color-gray-300;
+  border: 1px solid $color-gray-200;
   border-radius: 4px;
   display: flex;
+  justify-content: space-between;
+  flex-shrink: 0;
   align-items: center;
-  @include typo-c2;
+  @include typo-c1;
+
+  svg {
+    color: $color-gray-200;
+    margin-left: auto;
+  }
 }
 
 .trigger {
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  color: $color-gray-400;
-  background-color: #fff;
+  color: $color-gray-800;
+  background-color: $color-white;
+
+  svg {
+    path {
+      fill: $color-gray-800;
+    }
+  }
+
+  &:hover {
+    border-color: $color-orange-500;
+  }
+
+  &.open {
+    border-color: $color-orange-500;
+    box-shadow: 0px 0px 2px 0px $color-orange-500;
+  }
+}
+
+.small {
+  width: $small-width;
+  height: $small-height;
+  padding: 3px 6px 3px 10px;
+}
+
+.large {
+  width: $large-width;
+  height: $large-height;
+  padding-left: 10.5px;
+  padding-right: 14.5px;
 }
 
 .options {
   list-style: none;
-  padding: 0;
+  height: fit-content;
   margin: 0;
-  background-color: #fff;
+  margin-top: 4px;
+  padding: 0;
+  border-radius: 4px;
+  background-color: $color-white;
+  box-shadow: 0px 1px 3px 0px rgba(0, 0, 0, 0.25);
+  overflow: hidden;
+
+  &.small {
+    width: $small-width;
+  }
+
+  &.large {
+    width: $large-width;
+  }
 }
 
 .option {
-  border-top-color: transparent;
+  width: 100%;
+  color: $color-gray-500;
   cursor: pointer;
+  border: none;
+  border-radius: 0;
+  user-select: none;
 
-  .icon {
-    color: $color-gray-300;
-    margin-left: auto;
+  &.selected {
+    color: $color-selected-text;
+  }
+
+  &:hover {
+    background-color: $color-orange-100;
   }
 
   .selected-icon {
     color: $color-gray-800;
-  }
-
-  &.selected {
-    background-color: $color-gray-300;
-  }
-
-  &:not(.selected):hover {
-    background-color: #f0f0f0;
   }
 }
