+++ shared/ui/dropdown/types.ts
@@ -1,4 +1,4 @@
-import { CSSProperties } from 'react'
+import { CSSProperties, ReactNode } from 'react'
 
 export interface DropdownOptionModel {
   value: string
@@ -7,7 +7,21 @@ export interface DropdownOptionModel {
 
 export type DropdownValueType = string | string[] | null
 
+export type DropdownSizeType = 'small' | 'large'
+
+export interface DropdownProps {
+  size?: DropdownSizeType
+  Trigger: ReactNode
+  value: DropdownValueType
+  onChange: (value: DropdownValueType) => void
+  isMultiple: boolean
+  containerStyle?: CSSProperties
+  labelStyle?: CSSProperties
+  children?: ReactNode
+}
+
 export interface DropdownItemProps {
+  size?: DropdownSizeType
   value: string
   label: string
   isSelected: boolean
