+++ shared/ui/select/index.tsx
@@ -1,13 +1,20 @@
 'use client'
 
-// TODO: muliple 일 때는 상단에 placeholder를 그대로 둘지 고민
 import { CSSProperties } from 'react'
 
-import Dropdown, { DropdownProps } from '../dropdown'
-import { DropdownItemProps, DropdownOptionModel, DropdownValueType } from '../dropdown/types'
+import Dropdown from '../dropdown'
+import {
+  DropdownItemProps,
+  DropdownOptionModel,
+  DropdownProps,
+  DropdownValueType,
+} from '../dropdown/types'
 import { useSelect } from './hooks/useSelect'
 
-export type SelectType = Pick<DropdownProps, 'isMultiple' | 'containerStyle' | 'labelStyle'> &
+export type SelectType = Pick<
+  DropdownProps,
+  'size' | 'isMultiple' | 'containerStyle' | 'labelStyle'
+> &
   Pick<DropdownItemProps, 'hasCheck'> & {
     placeholder?: string
     value: DropdownValueType
@@ -18,6 +25,7 @@ export type SelectType = Pick<DropdownProps, 'isMultiple' | 'containerStyle' | '
   }
 
 const Select = ({
+  size,
   isRounded = false,
   isMultiple = false,
   hasCheck = false,
@@ -37,6 +45,7 @@ const Select = ({
 
   return (
     <Dropdown
+      size={size}
       Trigger={<span>{findLabel(options, value) ?? placeholder ?? options[0].label}</span>}
       isMultiple={isMultiple}
       value={value}
